Improve check of services, Fix #19
Check of services is now handled by the host using a small PHP script
that receives the URL to connect to and checks not only if it is
alive, but also checks its status code or returns 502 as fallback.
Check the status code is necessary in many cases, especially related
to the container world, since the status code can be used to see if a
service is really alive or dead (playing with healthcheckers and small
services that perform this task, for example).

The checkOnline() function in worker.js has been modified a bit for
reuse. I think another approach to implement this feature is to send
an identifier to the PHP script that does a read to settings.json to
get the URL of the website to check. I think this is a bit more complex,
but it only makes sense if the settings.json cannot be modified through
the web UI.